phase4_autogen/requirements.txt
# ============================================
# Phase 4: AutoGen - 依赖包
# ============================================
# 安装: pip install -r requirements.txt

# ============================================
# AutoGen 核心
# ============================================
pyautogen==0.2.28

# ============================================
# OpenAI 集成
# ============================================
openai==1.10.0

# ============================================
# 其他依赖
# ============================================
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.29.2

# ============================================
# 开发工具
# ============================================
black==23.12.1
isort==5.13.2
